requirements.txt
docker
matplotlib
pyyaml
latex
jinja2